Remove "Responsive to bug reports and feature requests"
I'm removing the "Responsive to bug reports and feature requests" line
from the chart because I'm no longer convinced that it's entirely
accurate; it's almost entirely subjective; and there's no good way to
maintain its accuracy over time.

Frankly I think I originally put it there to highlight how responsive
the Bitwarden team is to bug reports and feature requests, but since
then it has become clear that even Bitwarden is not consistent about
that. For example, they still haven't implemented unsharing or
duplicating items even though people have been asking them for those
features for over a year and the fact that they don't exist is a
serious deficiency in the product.